[#180] Use repository.type instead of Repository.scm_type throughout the plugin.
Here we have a nice inconsistency. scm_name is the same as the class name,
except for the abstract root class and "CVS". Here the class name is Cvs while
scm_name is "CVS". This inconsistency breaks the display for CVS repos.

Migrations ran fine, as I always used type there already.



git-svn-id: http://dev.holgerjust.de/svn/redmine-checkout/trunk@192 2c10314b-d60d-4f8f-b362-a6b3e82123ed