Triaging 'joe@dot.com' username bug
If a redmine username contains an '@' symbol (e.g. email address as username)
and "Display Login" is selected for a given protocol, that user will be
unable to view the "repository" page, and the server will return a 500 error.
Rather than debate how best to work around @ signs in usernames, this commit
declines to display login in checkout urls if the login name contains an '@'
symbol.